# src/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"version": "1.0.0",
		"title": "PEAT API",
		"description": "API para o app de agendamento de serviços em petshop PEAT"
	},
	"paths": {
...
		"/estabelecimento/{id}/imagens": {
			"post": {
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"pictures": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				},
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string",
							"minimum": 1
						},
						"description": "O id do estabelecimento"
					}
				],
				"summary": "Adiciona uma ou mais imagens do estabelecimento",
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"tags": ["Estabelecimento"]
			}
		},
		"/estabelecimento/{id}/avaliacao": {
			"post": {
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"avaliacao": {
										"type": "number",
										"example": 4
									}
								}
							}
						}
					}
				},
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string",
							"minimum": 1
						},
						"description": "O id do estabelecimento"
					}
				],
				"summary": "Adiciona uma avaliação ao estabelecimento",
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"tags": ["Estabelecimento"]
			}
		},
...
	},
	"components": {
...
		"securitySchemes": {
			"firebase": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "www.google.com",
						"scopes": {}
					}
				},
				"x-google-issuer": "https://securetoken.google.com/peat-f2f94",
				"x-google-jwks_uri": "https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com",
				"x-google-audiences": "peat-f2f94"
			}
		}
	}
}